Addressing review issues
