+++ regression/python/compound-assign/main.py
@@ -3,4 +3,3 @@ count *= 2
 count /= 2
 count += 1
 count -= 1
-result : int = 1/1
