+++ src/python-frontend/python_language.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "util/language.h"
+#include <util/language.h>
 
 #include <nlohmann/json.hpp>
